(fix)Calculator, Parser, Validator 분리 / (fix) Input,Output 메서드명 수정 / (fix) 입출력 분리 / (fix) 결과 저장 객체 Map 객체로 변경

+++ src/main/java/com/programmers/java/data/Result.java
@@ -5,8 +5,7 @@ import lombok.Getter;
 
 @AllArgsConstructor
 @Getter
-
 public class Result {
     private String exp;
-    private int answer;
+    private Double answer;
 }
